+++ src/swap/central/swapuz.ts
@@ -101,6 +101,7 @@ const MAINNET_CODE_TRANSCRIPTION: CurrencyPluginIdSwapChainCodeMap = {
   fio: null,
   groestlcoin: null,
   hedera: 'HBAR',
+  hyperevm: null,
   liberland: null,
   litecoin: 'LTC',
   monero: 'XMR',