+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+- added: Support for hyperevm to ChangeHero, Exolix, LetsExchange.
+- added: Support for hyperevm to Li.FI.
+- fixed: Fixed support for Metis on LI.FI.
+
 ## 2.28.0 (2025-07-17)
 
 - changed: Remove Thorchain minimum amounts from the thrown `SwapBelowLimit` errors